backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
python-decouple>=3.8
drf-spectacular>=0.27
django-filter>=23.5
Pillow>=10.0
gunicorn>=21.0

# Development
django-extensions>=3.2
ipython>=8.0

# Testing
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3
